Splitted engine.py and updated test cases for issue #93
