+++ src/datajunction/api/metrics.py
@@ -10,7 +10,7 @@ from sqlmodel import Session, SQLModel, select
 
 from datajunction.api.queries import save_query_and_run
 from datajunction.config import Settings
-from datajunction.engine import get_query_for_node
+from datajunction.sql.build import get_query_for_node
 from datajunction.models.node import Node, NodeType
 from datajunction.models.query import QueryWithResults
 from datajunction.utils import get_session, get_settings
